+++ NEWS.md
@@ -8,6 +8,9 @@ New language features
 Language changes
 ----------------
 
+* Converting arbitrary tuples to `NTuple`, e.g. `convert(NTuple, (1, ""))` now gives an error,
+  where it used to be incorrectly allowed. This is because `NTuple` refers only to homogeneous
+  tuples (this meaning has not changed) ([#34272]).
 
 Multi-threading changes
 -----------------------